Make "git-log --" without paths behave the same as "git-log" without --
"git log" family of commands, even when run from a subdirectory,
do not limit the revision range with the current directory as
the path limiter, but with double-dash without any paths after
it, i.e. "git log --" do so.  It was a mistake to have a
difference between "git log --" and "git log" introduced in
commit 4ab697aebb122c0f9890e3f4fc4f0f9b3a85f7cf (First cut at
libifying revlist generation).

Signed-off-by: Junio C Hamano <gitster@pobox.com>