Avoid scary errors about tagged trees/blobs during git-fetch
This is the same bug as d2ebc4c3f9203c5b2c2846fd14854d568ef1da77.
The warning "Object $X is a tree, not a commit" is bogus and is
not relevant here.  If its not a commit we just need to make sure
we don't mark it for merge as we fill out FETCH_HEAD.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>